Updated MAME core to MAME 0.71u3

+++ src/vidhrdw/psx.c
@@ -649,7 +649,7 @@ static void FlatPolygon( int n_points )
 			n_dx2 = (INT32)( ( COORD_X( m_packet.FlatPolygon.vertex[ n_rightpoint ].n_coord ) << 16 ) - n_cx2.d ) / n_distance;
 		}
 		n_x = n_cx1.w.h;
-		n_distance = n_cx2.w.h - n_x;
+		n_distance = (INT16)n_cx2.w.h - n_x;
 		if( n_distance > 0 && n_y >= m_n_drawarea_y1 && n_y <= m_n_drawarea_y2 )
 		{
 			if( (INT32)( m_n_drawarea_x1 - n_x ) > 0 )
@@ -913,7 +913,7 @@ static void FlatTexturedPolygon( int n_points )
 			n_dv2 = (INT32)( ( TEXTURE_V( m_packet.FlatTexturedPolygon.vertex[ n_rightpoint ].n_texture ) << 16 ) - n_cv2.d ) / n_distance;
 		}
 		n_x = n_cx1.w.h;
-		n_distance = n_cx2.w.h - n_x;
+		n_distance = (INT16)n_cx2.w.h - n_x;
 		if( n_distance > 0 && n_y >= m_n_drawarea_y1 && n_y <= m_n_drawarea_y2 )
 		{
 			n_u.d = n_cu1.d;
@@ -1245,7 +1245,7 @@ static void GouraudPolygon( int n_points )
 			n_db2 = (INT32)( ( BGR_B( m_packet.GouraudPolygon.vertex[ n_rightpoint ].n_bgr ) << 16 ) - n_cb2.d ) / n_distance;
 		}
 		n_x = n_cx1.w.h;
-		n_distance = n_cx2.w.h - n_x;
+		n_distance = (INT16)n_cx2.w.h - n_x;
 		if( n_distance > 0 && n_y >= m_n_drawarea_y1 && n_y <= m_n_drawarea_y2 )
 		{
 			n_r.d = n_cr1.d;
